BatteryService/requirements.txt
flask
flask_restful
cloudevents
requests
mysql-connector-python